+++ client/src/assets/images/admin/selection-background.png
