+++ client/src/utils/constants.js
@@ -1,0 +1,21 @@
+const FEATURES = [
+  'Skin Color',
+  'Back Hair',
+  'Front Hair',
+  'Side Hair',
+  'Extra Hair',
+  'Eyes',
+  'Eyebrows',
+  'Nose',
+  'Mouth',
+  'Blush',
+  'Accessories',
+];
+
+const LOGIN_FALLBACK = '/login';
+const ADMIN_FALLBACK = '/admin';
+const CUSTOMER_FALLBACK = '/create-sticker';
+
+export {
+  FEATURES, LOGIN_FALLBACK, ADMIN_FALLBACK, CUSTOMER_FALLBACK,
+};
